Changed the way the library works
I changed the way the library works. Instead of just feeding a number
into a function to show on the displays the library now gives the user
individual control of each display so they can set the number and
decimal on each display. I have tested it on one seven segment and it
works. I still have to test it on multiple and create documentation.